+++ code/src/cljs/sixsq/nuvla/ui/apps_store/events.cljs
@@ -7,9 +7,7 @@
             [sixsq.nuvla.ui.main.spec :as main-spec]
             [sixsq.nuvla.ui.plugins.full-text-search :as full-text-search-plugin]
             [sixsq.nuvla.ui.plugins.pagination :as pagination-plugin]
-            [sixsq.nuvla.ui.plugins.tab :as tab-plugin]
             [sixsq.nuvla.ui.session.spec :as session-spec]
-            [sixsq.nuvla.ui.session.utils :refer [get-active-claim]]
             [sixsq.nuvla.ui.utils.general :as general-utils]))
 
 (reg-event-fx
@@ -17,8 +15,7 @@
   (fn [{db :db}]
     {:db (merge db spec/defaults)
      :fx [[:dispatch [::apps-events/reset-version]]
-          [:dispatch [::apps-events/module-not-found false]]
-          [:dispatch [::get-modules]]]}))
+          [:dispatch [::apps-events/module-not-found false]]]}))
 
 (reg-event-db
   ::set-modules
@@ -29,26 +26,26 @@
 
 (reg-event-fx
   ::get-modules
-  (fn [{{:keys [::session-spec/session
-                ::spec/tab] :as db} :db}]
-    (-> {:db (assoc db ::apps-spec/module nil)
-         ::cimi-api-fx/search
-         [:module
-          (->> {:orderby "created:desc"
-                :filter  (general-utils/join-and
-                           (general-utils/join-or
-                             "subtype='component'"
-                             "subtype='application'"
-                             "subtype='application_kubernetes'")
-                           (case (::tab-plugin/active-tab tab)
-                             :appstore (general-utils/published-query-string)
-                             :myapps (general-utils/owner-like-query-string
-                                       (get-active-claim session))
-                             nil)
-                           (full-text-search-plugin/filter-text
-                             db [::spec/modules-search]))}
-               (pagination-plugin/first-last-params db [::spec/pagination]))
-          #(dispatch [::set-modules %])]})))
+  (fn [{{:keys [::session-spec/session] :as db} :db} [_ active-tab]]
+      (-> {:db (assoc db ::apps-spec/module nil)
+           ::cimi-api-fx/search
+           [:module
+            (->> {:orderby "created:desc"
+                  :filter  (general-utils/join-and
+                             (general-utils/join-or
+                               "subtype='component'"
+                               "subtype='application'"
+                               "subtype='application_kubernetes'")
+                             (case active-tab
+                               :appstore (general-utils/published-query-string)
+                               :myapps (general-utils/owner-like-query-string
+                                         (or (:active-claim session)
+                                           (:user session)))
+                               nil)
+                             (full-text-search-plugin/filter-text
+                               db [::spec/modules-search]))}
+                 (pagination-plugin/first-last-params db [(spec/page-keys->pagination-db-path active-tab)]))
+            #(dispatch [::set-modules %])]})))
 
 (reg-event-fx
   ::get-modules-summary