+++ code/src/cljs/sixsq/nuvla/ui/deployments_detail/views.cljs
@@ -20,7 +20,7 @@
             [sixsq.nuvla.ui.main.components :as components]
             [sixsq.nuvla.ui.main.events :as main-events]
             [sixsq.nuvla.ui.plugins.events :as events-plugin]
-            [sixsq.nuvla.ui.plugins.tab :as tab-plugin]
+            [sixsq.nuvla.ui.plugins.nav-tab :as tab-plugin]
             [sixsq.nuvla.ui.resource-log.views :as log-views]
             [sixsq.nuvla.ui.routing.events :as routing-events]
             [sixsq.nuvla.ui.routing.routes :as routes]
@@ -163,7 +163,7 @@
                 :icon    "linkify"}
      :render   #(r/as-element
                   [ui/TabPane [views-versions/versions-table
-                               @module-versions @module-content-id]])}))
+                               module-versions module-content-id]])}))
 
 
 (defn item-to-row